+++ packages/feeds-client/__integration-tests__/activity-added-event-filter.test.ts
@@ -26,9 +26,7 @@ describe('Feeds - onNewActivity filters WS events', () => {
   it('create feed and provide onNewActivity callback', async () => {
     feed = client.feed('user', crypto.randomUUID(), {
       onNewActivity: ({ activity }) =>
-        activity.filter_tags?.includes('important')
-          ? 'add-to-start'
-          : 'ignore',
+        activity.filter_tags?.includes('important') ? 'add-to-start' : 'ignore',
     });
     await feed.getOrCreate({
       watch: true,
@@ -37,19 +35,18 @@ describe('Feeds - onNewActivity filters WS events', () => {
       },
     });
 
+    const notImportantAdded = waitForEvent(feed, 'feeds.activity.added');
     serverClient.feeds.addActivity({
       type: 'post',
       feeds: [feed.feed],
       text: 'Not important post',
       user_id: user.id,
     });
-
-    await waitForEvent(feed, 'feeds.activity.added', {
-      shouldReject: true,
-    });
+    await notImportantAdded;
 
     expect(feed.state.getLatestValue().activities).toHaveLength(0);
 
+    const importantAdded = waitForEvent(feed, 'feeds.activity.added');
     serverClient.feeds.addActivity({
       type: 'post',
       feeds: [feed.feed],
@@ -57,10 +54,7 @@ describe('Feeds - onNewActivity filters WS events', () => {
       filter_tags: ['important'],
       user_id: user.id,
     });
-
-    await waitForEvent(feed, 'feeds.activity.added', {
-      shouldReject: true,
-    });
+    await importantAdded;
 
     expect(feed.state.getLatestValue().activities).toHaveLength(1);
     expect(feed.state.getLatestValue().activities?.[0].text).toBe(