+++ packages/feeds-client/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.3.0](https://github.com/GetStream/stream-feeds-js/compare/@stream-io/feeds-client-1.2.0...@stream-io/feeds-client-1.3.0) (2026-03-20)
+
+
+### Features
+
+* [FEEDS-1339] restore comments ([#268](https://github.com/GetStream/stream-feeds-js/issues/268)) ([e32d6fb](https://github.com/GetStream/stream-feeds-js/commit/e32d6fbe2d53d0d328333b370663f761546994fe))
+
+## [1.2.0](https://github.com/GetStream/stream-feeds-js/compare/@stream-io/feeds-client-1.1.0...@stream-io/feeds-client-1.2.0) (2026-03-19)
+
+
+### Features
+
+* support for flat notifications ([#267](https://github.com/GetStream/stream-feeds-js/issues/267)) ([e27a688](https://github.com/GetStream/stream-feeds-js/commit/e27a6885ed8754a831a311b86893c31d13849c01))
+
+
+### Bug Fixes
+
+* ensure that if enrich_own_fields are set to true own_* fields are updated even if HTTP response arrives after WS ([#262](https://github.com/GetStream/stream-feeds-js/issues/262)) ([94ebf6b](https://github.com/GetStream/stream-feeds-js/commit/94ebf6b0ccd0ddfe452d777a045899626c3b0273))
+
 ## [1.1.0](https://github.com/GetStream/stream-feeds-js/compare/@stream-io/feeds-client-1.0.0...@stream-io/feeds-client-1.1.0) (2026-03-10)